Initialize components returned by actions
This is a convenience hack. If the user doesn't use
pageWithName() to create a new component he wants to
return from an action, but rather does this:

    let nextPage = MyCoolPage()

i.e. he is just allocating the page, the page won't
be properly initialized (no template assigned, no
context, no subcomponents, etc).

This patch adjusts the components so that this works.